+++ packages/react-client/src/devices/mediaInitializer.ts
@@ -1,4 +1,4 @@
-import { prepareConstraints } from "../constraints";
+import { prepareConstraints } from "./constraints";
 import type { CurrentDevices, DeviceError } from "../types/internal";
 import { NOT_FOUND_ERROR, OVERCONSTRAINED_ERROR, PERMISSION_DENIED, UNHANDLED_ERROR } from "../utils/errors";
